Move some test files to Java6 file naming
